refactor item schema to separate file
